requirements.txt
aiohappyeyeballs
aiohttp
aiosignal
annotated-types
anyio
attrs
bitarray
bitstring
blinker
certifi
cffi
charset-normalizer
click
colorama
cryptography
dataclasses-json
decorator
distro
esptool
et_xmlfile
fade
faiss-cpu
filetype
Flask
flask-cors
frozenlist
google-ai-generativelanguage
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-genai
google-generativeai
googleapis-common-protos
greenlet
grpcio
grpcio-status
h11
httpcore
httplib2
httpx
httpx-sse
idna
intelhex
itsdangerous
Jinja2
jsonpatch
jsonpointer
langchain
langchain-classic
langchain-community
langchain-core
langchain-google-genai
langchain-text-splitters
langgraph
langgraph-checkpoint
langgraph-prebuilt
langgraph-sdk
langsmith
lxml
markdown-it-py
MarkupSafe
marshmallow
mdurl
MouseInfo
multidict
mypy_extensions
numpy
openpyxl
orjson
ormsgpack
packaging
pandas
pillow
propcache
proto-plus
protobuf
psutil
pyasn1
pyasn1_modules
PyAutoGUI
pycparser
pydantic
pydantic-settings
pydantic_core
PyGetWindow
Pygments
PyMsgBox
pyparsing
pypdf
pyperclip
PyRect
PyScreeze
pyserial
pyte
python-dateutil
python-docx
python-dotenv
python-pptx
pytweening
pytz
PyYAML
reedsolo
requests
requests-toolbelt
rich
rich-click
setuptools
six
sniffio
SQLAlchemy
tenacity
thefuck
tqdm
typing-inspect
typing-inspection
typing_extensions
tzdata
uritemplate
urllib3
uuid_utils
wcwidth
websockets
Werkzeug
win_unicode_console
xlsxwriter
xxhash
yarl
zstandard
